sorted connectWallet && working on create election
