+++ frontend/src/app/layout.tsx
@@ -4,6 +4,7 @@ import { StarknetProvider } from "@/components/starknet-provider";
 import "./globals.css";
 import Navbar from "../components/Navbar";
 import Footer from "../components/Footer";
+import { WalletProvider } from "../context/WalletContext";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -21,13 +22,15 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className}>
         <StarknetProvider>
-          <div className=" text-[#000]">
-            <Navbar />
-            <div className="pt-24 text-[#000] bg-[#f7ebeb] pb-24">
-              <main>{children}</main>
+          <WalletProvider>
+            <div className=" text-[#000]">
+              <Navbar />
+              <div className="pt-24 text-[#000] bg-[#f7ebeb] pb-24">
+                <main>{children}</main>
+              </div>
+              <Footer />
             </div>
-            <Footer />
-          </div>
+          </WalletProvider>
         </StarknetProvider>
       </body>
     </html>
